+++ sdk/communication/Azure.Communication.CallAutomation/src/CallAutomationClient.cs
@@ -821,7 +821,7 @@ namespace Azure.Communication.CallAutomation
                 ? default
                 : new MediaStreamingOptionsInternal(configuration.TransportUri.AbsoluteUri, configuration.MediaStreamingTransport,
                 configuration.MediaStreamingContent, configuration.MediaStreamingAudioChannel, configuration.StartMediaStreaming,
-                configuration.EnableBidirectional, configuration.AudioFormat == null ? AudioFormat.Pcm24KMono : configuration.AudioFormat);
+                configuration.EnableBidirectional, configuration.AudioFormat);
         }
         private static TranscriptionOptionsInternal CreateTranscriptionOptionsInternal(TranscriptionOptions configuration)
         {