+++ sdk/storage/Azure.Storage.DataMovement.Blobs/src/BlobDestinationCheckpointDetails.cs
@@ -18,31 +18,30 @@ namespace Azure.Storage.DataMovement.Blobs
         /// <summary>
         /// The type of blob.
         /// </summary>
-        public DataTransferProperty<BlobType?> BlobType;
-        public bool PreserveBlobType;
-        public BlobType? BlobTypeValue;
+        public bool IsBlobTypeSet;
+        public BlobType? BlobType;
 
         /// <summary>
         /// The content headers for the destination blob.
         /// </summary>
-        public DataTransferProperty<string> CacheControl;
-        public bool PreserveCacheControl;
+        public string CacheControl;
+        public bool IsCacheControlSet;
         public byte[] CacheControlBytes;
 
-        public DataTransferProperty<string> ContentDisposition;
-        public bool PreserveContentDisposition;
+        public string ContentDisposition;
+        public bool IsContentDispositionSet;
         public byte[] ContentDispositionBytes;
 
-        public DataTransferProperty<string> ContentEncoding;
-        public bool PreserveContentEncoding;
+        public string ContentEncoding;
+        public bool IsContentEncodingSet;
         public byte[] ContentEncodingBytes;
 
-        public DataTransferProperty<string> ContentLanguage;
-        public bool PreserveContentLanguage;
+        public string ContentLanguage;
+        public bool IsContentLanguageSet;
         public byte[] ContentLanguageBytes;
 
-        public DataTransferProperty<string> ContentType;
-        public bool PreserveContentType;
+        public string ContentType;
+        public bool IsContentTypeSet;
         public byte[] ContentTypeBytes;
 
         /// <summary>
@@ -53,64 +52,98 @@ namespace Azure.Storage.DataMovement.Blobs
         /// <summary>
         /// The metadata for the destination blob.
         /// </summary>
-        public DataTransferProperty<Metadata> Metadata;
-        public bool PreserveMetadata;
+        public Metadata Metadata;
+        public bool IsMetadataSet;
         public byte[] MetadataBytes;
 
         /// <summary>
         /// The Blob tags for the destination blob.
         /// </summary>
-        public DataTransferProperty<Tags> Tags;
+        public Tags Tags;
         public bool PreserveTags;
         public byte[] TagsBytes;
 
         public override int Length => CalculateLength();
 
+        public BlobDestinationCheckpointDetails(BlobStorageResourceContainerOptions options) : this(
+            isBlobTypeSet: options?._isBlobTypeSet ?? false,
+            blobType: options?.BlobType,
+            blobOptions: options?.BlobOptions)
+        { }
+
+        public BlobDestinationCheckpointDetails(bool isBlobTypeSet, BlobType? blobType, BlobStorageResourceOptions blobOptions)
+            : this(
+                isBlobTypeSet: isBlobTypeSet,
+                blobType: blobType,
+                isContentTypeSet: blobOptions?._isContentTypeSet ?? false,
+                contentType: blobOptions?.ContentType,
+                isContentEncodingSet: blobOptions?._isContentEncodingSet ?? false,
+                contentEncoding: blobOptions?.ContentEncoding,
+                isContentLanguageSet: blobOptions?._isContentLanguageSet ?? false,
+                contentLanguage: blobOptions?.ContentLanguage,
+                isContentDispositionSet: blobOptions?._isContentDispositionSet ?? false,
+                contentDisposition: blobOptions?.ContentDisposition,
+                isCacheControlSet: blobOptions?._isCacheControlSet ?? false,
+                cacheControl: blobOptions?.CacheControl,
+                accessTier: blobOptions?.AccessTier,
+                isMetadataSet: blobOptions?._isMetadataSet ?? false,
+                metadata: blobOptions?.Metadata,
+                preserveTags: true,
+                tags: default)
+        { }
+
         public BlobDestinationCheckpointDetails(
-            DataTransferProperty<BlobType?> blobType,
-            DataTransferProperty<string> contentType,
-            DataTransferProperty<string> contentEncoding,
-            DataTransferProperty<string> contentLanguage,
-            DataTransferProperty<string> contentDisposition,
-            DataTransferProperty<string> cacheControl,
+            bool isBlobTypeSet,
+            BlobType? blobType,
+            bool isContentTypeSet,
+            string contentType,
+            bool isContentEncodingSet,
+            string contentEncoding,
+            bool isContentLanguageSet,
+            string contentLanguage,
+            bool isContentDispositionSet,
+            string contentDisposition,
+            bool isCacheControlSet,
+            string cacheControl,
             AccessTier? accessTier,
-            DataTransferProperty<Metadata> metadata,
-            DataTransferProperty<Tags> tags)
+            bool isMetadataSet,
+            Metadata metadata,
+            bool preserveTags,
+            Tags tags)
         {
             Version = DataMovementBlobConstants.DestinationCheckpointDetails.SchemaVersion;
             BlobType = blobType;
-            PreserveBlobType = blobType?.Preserve ?? true;
-            BlobTypeValue = blobType?.Value != default ? blobType.Value : default;
+            IsBlobTypeSet = isBlobTypeSet;
 
             AccessTierValue = accessTier;
 
             CacheControl = cacheControl;
-            PreserveCacheControl = cacheControl?.Preserve ?? true;
-            CacheControlBytes = cacheControl?.Value != default ? Encoding.UTF8.GetBytes(cacheControl.Value) : Array.Empty<byte>();
+            IsCacheControlSet = isCacheControlSet;
+            CacheControlBytes = cacheControl != default ? Encoding.UTF8.GetBytes(cacheControl) : Array.Empty<byte>();
 
             ContentDisposition = contentDisposition;
-            PreserveContentDisposition = contentDisposition?.Preserve ?? true;
-            ContentDispositionBytes = contentDisposition?.Value != default ? Encoding.UTF8.GetBytes(contentDisposition.Value) : Array.Empty<byte>();
+            IsContentDispositionSet = isContentDispositionSet;
+            ContentDispositionBytes = contentDisposition != default ? Encoding.UTF8.GetBytes(contentDisposition) : Array.Empty<byte>();
 
             ContentEncoding = contentEncoding;
-            PreserveContentEncoding = contentEncoding?.Preserve ?? true;
-            ContentEncodingBytes = contentEncoding?.Value != default ? Encoding.UTF8.GetBytes(contentEncoding.Value) : Array.Empty<byte>();
+            IsContentEncodingSet = isContentEncodingSet;
+            ContentEncodingBytes = contentEncoding!= default ? Encoding.UTF8.GetBytes(contentEncoding) : Array.Empty<byte>();
 
             ContentLanguage = contentLanguage;
-            PreserveContentLanguage = contentLanguage?.Preserve ?? true;
-            ContentLanguageBytes = contentLanguage?.Value != default ? Encoding.UTF8.GetBytes(contentLanguage.Value) : Array.Empty<byte>();
+            IsContentLanguageSet = isContentLanguageSet;
+            ContentLanguageBytes = contentLanguage != default ? Encoding.UTF8.GetBytes(contentLanguage) : Array.Empty<byte>();
 
             ContentType = contentType;
-            PreserveContentType = contentType?.Preserve ?? true;
-            ContentTypeBytes = contentType?.Value != default ? Encoding.UTF8.GetBytes(contentType.Value) : Array.Empty<byte>();
+            IsContentTypeSet = isContentTypeSet;
+            ContentTypeBytes = contentType != default ? Encoding.UTF8.GetBytes(contentType) : Array.Empty<byte>();
 
             Metadata = metadata;
-            PreserveMetadata = metadata?.Preserve ?? true;
-            MetadataBytes = metadata?.Value != default ? Encoding.UTF8.GetBytes(metadata.Value.DictionaryToString()) : Array.Empty<byte>();
+            IsMetadataSet = isMetadataSet;
+            MetadataBytes = metadata != default ? Encoding.UTF8.GetBytes(metadata.DictionaryToString()) : Array.Empty<byte>();
 
             Tags = tags;
-            PreserveTags = tags?.Preserve ?? false;
-            TagsBytes = tags?.Value != default ? Encoding.UTF8.GetBytes(tags.Value.DictionaryToString()) : Array.Empty<byte>();
+            PreserveTags = preserveTags;
+            TagsBytes = tags != default ? Encoding.UTF8.GetBytes(tags.DictionaryToString()) : Array.Empty<byte>();
         }
 
         protected override void Serialize(Stream stream)
@@ -124,10 +157,10 @@ namespace Azure.Storage.DataMovement.Blobs
             writer.Write(Version);
 
             // BlobType
-            writer.Write(PreserveBlobType);
-            if (!PreserveBlobType)
+            writer.Write(IsBlobTypeSet);
+            if (IsBlobTypeSet)
             {
-                writer.Write((byte)BlobTypeValue);
+                writer.Write((byte)BlobType);
             }
             else
             {
@@ -135,8 +168,8 @@ namespace Azure.Storage.DataMovement.Blobs
             }
 
             // Preserve Content Type
-            writer.Write(PreserveContentType);
-            if (!PreserveContentType)
+            writer.Write(IsContentTypeSet);
+            if (IsContentTypeSet)
             {
                 // Content Type offset/length
                 writer.WriteVariableLengthFieldInfo(ContentTypeBytes.Length, ref currentVariableLengthIndex);
@@ -148,8 +181,8 @@ namespace Azure.Storage.DataMovement.Blobs
             }
 
             // Preserve Content Encoding
-            writer.Write(PreserveContentEncoding);
-            if (!PreserveContentEncoding)
+            writer.Write(IsContentEncodingSet);
+            if (IsContentEncodingSet)
             {
                 // ContentEncoding offset/length
                 writer.WriteVariableLengthFieldInfo(ContentEncodingBytes.Length, ref currentVariableLengthIndex);
@@ -161,8 +194,8 @@ namespace Azure.Storage.DataMovement.Blobs
             }
 
             // Preserve Content Language
-            writer.Write(PreserveContentLanguage);
-            if (!PreserveContentLanguage)
+            writer.Write(IsContentLanguageSet);
+            if (IsContentLanguageSet)
             {
                 // ContentLanguage offset/length
                 writer.WriteVariableLengthFieldInfo(ContentLanguageBytes.Length, ref currentVariableLengthIndex);
@@ -174,8 +207,8 @@ namespace Azure.Storage.DataMovement.Blobs
             }
 
             // Preserve Content Disposition
-            writer.Write(PreserveContentDisposition);
-            if (!PreserveContentDisposition)
+            writer.Write(IsContentDispositionSet);
+            if (IsContentDispositionSet)
             {
                 // ContentDisposition offset/length
                 writer.WriteVariableLengthFieldInfo(ContentDispositionBytes.Length, ref currentVariableLengthIndex);
@@ -187,8 +220,8 @@ namespace Azure.Storage.DataMovement.Blobs
             }
 
             // Preserve Cache Control
-            writer.Write(PreserveCacheControl);
-            if (!PreserveCacheControl)
+            writer.Write(IsCacheControlSet);
+            if (IsCacheControlSet)
             {
                 // CacheControl offset/length
                 writer.WriteVariableLengthFieldInfo(CacheControlBytes.Length, ref currentVariableLengthIndex);
@@ -203,8 +236,8 @@ namespace Azure.Storage.DataMovement.Blobs
             writer.Write((byte)AccessTierValue.ToJobPlanAccessTier());
 
             // Preserve Metadata
-            writer.Write(PreserveMetadata);
-            if (!PreserveMetadata)
+            writer.Write(IsMetadataSet);
+            if (IsMetadataSet)
             {
                 // Metadata offset/length
                 writer.WriteVariableLengthFieldInfo(MetadataBytes.Length, ref currentVariableLengthIndex);
@@ -228,27 +261,27 @@ namespace Azure.Storage.DataMovement.Blobs
                 writer.WriteEmptyLengthOffset();
             }
 
-            if (!PreserveContentType)
+            if (IsContentTypeSet)
             {
                 writer.Write(ContentTypeBytes);
             }
-            if (!PreserveContentEncoding)
+            if (IsContentEncodingSet)
             {
                 writer.Write(ContentEncodingBytes);
             }
-            if (!PreserveContentLanguage)
+            if (IsContentLanguageSet)
             {
                 writer.Write(ContentLanguageBytes);
             }
-            if (!PreserveContentDisposition)
+            if (IsContentDispositionSet)
             {
                 writer.Write(ContentDispositionBytes);
             }
-            if (!PreserveCacheControl)
+            if (IsCacheControlSet)
             {
                 writer.Write(CacheControlBytes);
             }
-            if (!PreserveMetadata)
+            if (IsMetadataSet)
             {
                 writer.Write(MetadataBytes);
             }
@@ -273,31 +306,31 @@ namespace Azure.Storage.DataMovement.Blobs
 
             // Index Values
             // BlobType
-            bool preserveBlobType = reader.ReadBoolean();
+            bool isBlobTypeSet = reader.ReadBoolean();
             BlobType blobType = (BlobType)reader.ReadByte();
 
             // Preserve Content Type and offset/length
-            bool preserveContentType = reader.ReadBoolean();
+            bool isContentTypeSet = reader.ReadBoolean();
             int contentTypeOffset = reader.ReadInt32();
             int contentTypeLength = reader.ReadInt32();
 
             // Preserve Content Encoding and offset/length
-            bool preserveContentEncoding = reader.ReadBoolean();
+            bool isContentEncodingSet = reader.ReadBoolean();
             int contentEncodingOffset = reader.ReadInt32();
             int contentEncodingLength = reader.ReadInt32();
 
             // Preserve Content Language and offset/length
-            bool preserveContentLanguage = reader.ReadBoolean();
+            bool isContentLanguageSet = reader.ReadBoolean();
             int contentLanguageOffset = reader.ReadInt32();
             int contentLanguageLength = reader.ReadInt32();
 
             // Preserve ContentDisposition and offset/length
-            bool preserveContentDisposition = reader.ReadBoolean();
+            bool isContentDispositionSet = reader.ReadBoolean();
             int contentDispositionOffset = reader.ReadInt32();
             int contentDispositionLength = reader.ReadInt32();
 
             // Preserve CacheControl and offset/length
-            bool preserveCacheControl = reader.ReadBoolean();
+            bool isCacheControlSet = reader.ReadBoolean();
             int cacheControlOffset = reader.ReadInt32();
             int cacheControlLength = reader.ReadInt32();
 
@@ -310,7 +343,7 @@ namespace Azure.Storage.DataMovement.Blobs
             }
 
             // Preserve Metadata and offset/length
-            bool preserveMetadata = reader.ReadBoolean();
+            bool isMetadataSet = reader.ReadBoolean();
             int metadataOffset = reader.ReadInt32();
             int metadataLength = reader.ReadInt32();
 
@@ -377,15 +410,23 @@ namespace Azure.Storage.DataMovement.Blobs
             }
 
             return new BlobDestinationCheckpointDetails(
-                blobType: preserveBlobType ? new(preserveBlobType) : new(blobType),
-                contentType: preserveContentType ? new(preserveContentType) : new(contentType),
-                contentEncoding: preserveContentEncoding ? new(preserveContentEncoding): new(contentEncoding),
-                contentLanguage: preserveContentLanguage ? new(preserveContentLanguage) : new(contentLanguage),
-                contentDisposition: preserveContentDisposition ? new(preserveContentDisposition) : new(contentDisposition),
-                cacheControl: preserveCacheControl ? new(preserveCacheControl): new(cacheControl),
+                isBlobTypeSet: isBlobTypeSet,
+                blobType: blobType,
+                isContentTypeSet: isContentTypeSet,
+                contentType: contentType,
+                isContentEncodingSet: isContentEncodingSet,
+                contentEncoding: contentEncoding,
+                isContentLanguageSet: isContentLanguageSet,
+                contentLanguage: contentLanguage,
+                isContentDispositionSet: isContentDispositionSet,
+                contentDisposition: contentDisposition,
+                isCacheControlSet: isCacheControlSet,
+                cacheControl: cacheControl,
                 accessTier: accessTier,
-                metadata: preserveMetadata ? new(preserveMetadata) : new(metadataString.ToDictionary(nameof(metadataString))),
-                tags: preserveTags ? new(preserveTags) : new(tagsString.ToDictionary(nameof(tagsString))));
+                isMetadataSet: isMetadataSet,
+                metadata: metadataString.ToDictionary(nameof(metadataString)),
+                preserveTags: preserveTags,
+                tags: tagsString.ToDictionary(nameof(tagsString)));
         }
 
         private int CalculateLength()
@@ -393,27 +434,27 @@ namespace Azure.Storage.DataMovement.Blobs
             // Length is calculated based on whether the property is preserved.
             // If the property is preserved, the property's length is added to the total length.
             int length = DataMovementBlobConstants.DestinationCheckpointDetails.VariableLengthStartIndex;
-            if (!PreserveContentType)
+            if (!IsContentTypeSet)
             {
                 length += ContentTypeBytes.Length;
             }
-            if (!PreserveContentEncoding)
+            if (!IsContentEncodingSet)
             {
                 length += ContentEncodingBytes.Length;
             }
-            if (!PreserveContentLanguage)
+            if (!IsContentLanguageSet)
             {
                 length += ContentLanguageBytes.Length;
             }
-            if (!PreserveContentDisposition)
+            if (!IsContentDispositionSet)
             {
                 length += ContentDispositionBytes.Length;
             }
-            if (!PreserveCacheControl)
+            if (!IsCacheControlSet)
             {
                 length += CacheControlBytes.Length;
             }
-            if (!PreserveMetadata)
+            if (!IsMetadataSet)
             {
                 length += MetadataBytes.Length;
             }