+++ sdk/storage/Azure.Storage.DataMovement/src/LocalFilesStorageResourceProvider.cs
@@ -20,7 +20,7 @@ namespace Azure.Storage.DataMovement
         /// <summary>
         /// Default constructor.
         /// </summary>
-        public LocalFilesStorageResourceProvider()
+        internal LocalFilesStorageResourceProvider()
         {
         }
 
@@ -50,7 +50,7 @@ namespace Azure.Storage.DataMovement
         /// <returns>
         /// Storage resource to this file.
         /// </returns>
-        public StorageResourceItem FromFile(string filePath)
+        public static StorageResourceItem FromFile(string filePath)
         {
             return new LocalFileStorageResource(filePath);
         }
@@ -64,7 +64,7 @@ namespace Azure.Storage.DataMovement
         /// <returns>
         /// Storage resource to this directory.
         /// </returns>
-        public StorageResourceContainer FromDirectory(string directoryPath)
+        public static StorageResourceContainer FromDirectory(string directoryPath)
         {
             return new LocalDirectoryStorageResourceContainer(directoryPath);
         }