+++ sdk/storage/Azure.Storage.DataMovement/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 ### Breaking Changes
+- Removed `DataTransferProperty` and `DataTransferProperty<T>`
 - Renamed the following types/properties:
     - `DataTransfer` -> `TransferOperation`
         - Addtionally renamed the `TransferStatus` property to `Status`
@@ -16,6 +17,8 @@
     - `DataTransferState` -> `TransferState`
     - `DataTransferStatus` -> `TransferStatus`
     - `DataTransferErrorMode` -> `TransferErrorMode`
+    - `ProgressHandlerOptions` -> `TransferProgressHandlerOptions`
+        - Also removed the constructor since properties are settable.
     - `StorageResourceCheckpointData` -> `StorageResourceCheckpointDetails`
     - `StorageResource.GetDestinationCheckpointData` -> `StorageResource.GetDestinationCheckpointDetails`
     - `StorageResource.GetSourceCheckpointData` -> `StorageResource.GetSourceCheckpointDetails`
@@ -32,6 +35,10 @@
     - `TransferCheckpointStoreOptions.Disabled` -> `TransferCheckpointStoreOptions.DisableCheckpoint`
 - Removed properties from `StorageResourceItemProperties` constructor since properties are settable.
 - Changed type of `StorageResourceItemProperties.RawProperties` to `IDictionary`.
+- Changed `List<StorageResourceProvider> TransferManagerOptions.ResumeProviders` to `IList<StorageResourceProvider> TransferManagerOptions.ProvidersForResuming`
+- Changed the following `LocalFilesStorageResourceProvider` methods to `static` methods:
+    - `FromFile(string)`
+    - `FromDirectory(string)`
 
 ### Bugs Fixed
 - Fixed bug where adding multiple transfers in parallel could cause a collision (`InvalidOperationException`) in the data transfers stored within the `TransferManager`.